Added BoxExternal* classes
* BoxExternal is now only an interface with a get/set owner method
* BoxExternalFile extends BoxFile and implements BoxExternal
* BoxExternalFolder extends BoxFolder and implements BoxExternal
* BoxExternalReference holds information to a shared
Directory/FileMetadata file
