+++ qabelbox/src/main/java/de/qabel/qabelbox/storage/BoxExternalFolder.java
@@ -2,32 +2,12 @@ package de.qabel.qabelbox.storage;
 
 import de.qabel.core.crypto.QblECPublicKey;
 
-public class BoxExternalFile extends BoxFile implements BoxExternal {
+public class BoxExternalFolder extends BoxFolder implements BoxExternal {
 
 	public QblECPublicKey owner;
 
-	public BoxExternalFile(QblECPublicKey owner, String block, String name, Long size, Long mtime, byte[] key) {
-		super(block, name, size, mtime, key);
-		this.owner = owner;
-	}
-
-	@Override
-	public boolean equals(Object o) {
-		if (this == o) return true;
-		if (o == null || getClass() != o.getClass()) return false;
-		if (!super.equals(o)) return false;
-
-		BoxExternalFile that = (BoxExternalFile) o;
-
-		return !(owner != null ? !owner.equals(that.owner) : that.owner != null);
-
-	}
-
-	@Override
-	public int hashCode() {
-		int result = super.hashCode();
-		result = 31 * result + (owner != null ? owner.hashCode() : 0);
-		return result;
+	public BoxExternalFolder(String ref, String name, byte[] key) {
+		super(ref, name, key);
 	}
 
 	@Override
@@ -38,5 +18,24 @@ public class BoxExternalFile extends BoxFile implements BoxExternal {
 	@Override
 	public void setOwner(QblECPublicKey owner) {
 		this.owner = owner;
+	}
+
+	@Override
+	public boolean equals(Object o) {
+		if (this == o) return true;
+		if (o == null || getClass() != o.getClass()) return false;
+		if (!super.equals(o)) return false;
+
+		BoxExternalFolder that = (BoxExternalFolder) o;
+
+		return !(owner != null ? !owner.equals(that.owner) : that.owner != null);
+
+	}
+
+	@Override
+	public int hashCode() {
+		int result = super.hashCode();
+		result = 31 * result + (owner != null ? owner.hashCode() : 0);
+		return result;
 	}
 }
